Change some bugs regarding handling of top and bottom page margins.
If the printer head moves below the bottom margin, we're supposed to
eject the page and start a new one.  That's a little tricky with PPM,
especially when writing to stdout.  I think the best behavior on a
page ejection situation is to just write out the PPM with what has
been printed so far and just ignore the rest of the data.  In other
words, unprint will produce a PPM file describing the first page of
output for any given escp2 input.

The only other option would be to write out several PPMs in one file.  I
was hoping that "pnmcat" would be able to read multiple files from stdin,
but that seems not to be the case.