Fix the bugs that prevented softweave from working properly.
This is currently hard coded for the 740/750/1200 nozzle separation of
6/720".  If you want to run it on output for the printers with a nozzle
sep of 8/720, you need to change a line.

Ideally we need a command line switch to select the model, but, since it's
after 1AM that's not going to happen before I go home tonight.  At this point
we need to decide how much code we want to share between print and
unprint.  Since unprint is designed to be a check of print, one argument
is to keep it completely independent, but, that's more work....