Allow printing to the top edge and much closer to the bottom on Epson printers (major change in the softweave code).
Performance tweaks in the dither routines and in the color conversion
routine (this needs to be ported to the other conversion routines in
print-util.c).

Support for some kind of ink limiting in the dither code.  Not clear that
this really does very well (it may make things ugly); we may rip it back
out.