Tune ink values for 870.
Flip flop again on how to handle variable dots with error diffusion.
I believe that what I'm now doing is correct, but it needs a better matrix
in order to really work properly.  Hopefully Thomas Tonino will provide
that.

Use ordered dithering only for black ink, to escape error diffusion
artifacts.  This may not be a good decision; time will tell.  We also
need a better matrix for that.