Do a better job of duplicate line elimination in the dither code.  We now get notified by the driver if there are known duplicate lines. This would allow us to put something in the Image layer to optimize out blank lines very early on and save a lot of computation.
A few point optimizations in the dither code.

Move the lut_t into the private data of print_util; it isn't actually
needed anywhere else.  Also hang some stuff off it to optimize the
lookup for the photo case.

Get rid of some potentially dangerous FP equality comparisons in
the HSL code.  Also avoid the scaling to 1 that both endangers
the computation and wastes time.