Add a "strategy" argument to the weave code.  This is currently taken from the use_softweave member of the res_t list.  The purpose is to allow different softweave patterns to be used; depending on the circumstances, different weave might be appropriate.  (I don't suggest letting the user select a weave, but the driver might do this.)
At the moment, weave pattern 1 is identical to what we had before.  On my
870 in 720 softweave, pattern 4 produces less banding than pattern 1,
while patterns 2 and 3 seem to produce more.  Pattern 1's banding comes
in "waves".  Pattern 6 seems to do reasonably well...  I think this has
something to do with what happens when a row of ink is laid down touching
another row which isn't fully dry yet.  If this is the case, it's not
obvious how to decide which strategy to use in any given situation.