Updated print densities for 2000P; the previous ones were just a little low, so I bumped them by 20%.
Updated configure script and makefile to correctly figure out the right
configuration directory for CUPS.

Updated the makefile again to use .SILENT and the echos, but added a
set of echos at the beginning to show the compiler and options that are
being used.  This is much cleaner than a 3+ line listing for each source
file in a make log for those extra-long autoconf CFLAGS lines.

Fixed include file nonsense; all needed include files should be there now,
and all duplicates removed.