Fixes for PCL and BJC driver resolutions - they use XRESxYRES, while the EPSON driver uses "XRES x YRES"...
Added monochrome resolution mode for PCL driver.

Added mergevars and validatevars calls per Robert's patch.

The install target now depends on "all".

Added code to limit the number of lines read to the raster height, and
to flush any unused lines in the buffer.  This should only be needed if
the GIMP-print driver code gives up on printing before reading all of the
image data...