447865 print-weave addresses too many print heads
The problem here is that the weave code defines the head offset to be
an array of 8 offsets; everything else uses 7 heads (channels).  The
Lexmark driver was confusing two different things, but having a
different number of head offsets is a sure way to cause confusion.

Also don't offset the top when printing in black and white only on
ESC 480/580.  When printing with black-only ink, that printer has a
"normal" head.