Support the Stylus Scan 2500
Add minimum paper sizes for HP printers, from Dave Hill.  I'm checking
this in because this is needed to work around a CUPS bug (see
our bug 470240).  The symptoms are that when the number of paper
sizes exceeds 94 in the DesignJet 750 PPD file CUPS loses
track of all printers in the administrative interface.