I believe I've finally solved (at least in Gimp-Print) the problem of handling variable drop sizes.  The trick is in the interpretation of the error value.  The code change is quite straightforward.
The assumption is that we always want to print the larger or smaller
drop size (the smaller drop size may be zero) of the range defined by
the unadjusted, uncorrected input value.  This is because we always
want to fill the page with small drops before printing any large
drops(*).  We don't want to print unnecessary large drops or leave
unnecessary holes.

The insight is that at any given point we aren't trying to perform an
N-value dither; we're trying to perform an ordinary single-value
dither -- on or off, where "on" corresponds to the larger drop and
"off" corresponds to the smaller drop.  After all, if the page is
filled with small drops, we might as well treat them as though they're
turned off, since they're all the same anyway.  So the trick is that
we simply compute a new input value that's defined as the point within
the range, and use this value in all calculations.  The error value is
based on this relative value; the raw input is only used to pick the
range and compute where within the range we lie.

For example: suppose we have three drop sizes with values of 8000,
16000, and 65535 (on a scale of 0-65535), and an input value of 10000.
We use the input value of 10000 to determine that the choices we're
interested in are drop sizes 1 and 2.  We compute the new, relative
input value as 65535 * (10000 - 8000) / (16000 - 8000), or 16383.  So
we want to print 1/4 of the positions with drop size 2 and 3/4 with
drop size 1.

We now perform normal Even Tone dithering on this value of 16383.  So
if we decide to print a large drop, we have an error of 49152, and if
we don't print we have an error of -16383.  Simple!