dyesub: kodak1400/805: Fix margins so they display and function properly.
These printers require full-bleed output *horizontally* though the
imaging area is smaller than the data supplied.  Further complicating
things is that the lamination area is actually smaller than the
printable area.

So, set the margins so that we can't print outside the lamination area,
but ensure gutenprint generates the full-bleed width regardless.