Further shrink saved checksums by abbreviating the hash.
Hash is shortened to the shortest unambiguous prefix.  The full SHA-512
hash is 128 bytes; the hash is essentially incompressible beyond 64 bytes,
and takes up most of the space in the bzip2-compressed output.

The purpose of the hash is to allow us to efficiently determine whether
the output has changed between releases/snapshots.  A shortened hash
increases the probability that output will change without us detecting it.
If the minimum common prefix is 7 hex bytes (so we use 8 bytes of hash), the
likelihood that any given change will be undetected is 1 in 2^32, which is
small enough not to worry about.

An additional optimization is to store only the delta suffix from the
previous hash (with sorted hashes).

We compare only the first min(N,M) bytes of hash for files storing
M and N bytes respectively.

The savings with this is quite substantial; with 5.2.12-pre4 it's about 80%
over the previous compression scheme, which is about 60% over bzip2 of the
raw output.  This makes it practical for us to run and archive test
results much more frequently.