dyesub:  Simplify dyesub_print_pixel.
As all current printer and color models map 1:1 from image->inks, get
rid of the useless (not to mention broken!) code that allowed the dyesub
driver to handle mismatches.

This also includes the initial work towards a better interpolation
algorithm -- Right now the driver just picks the closest single "point"
in the original image and uses that as-is for a given output pixel.

Eventually, I want to allow for different scaling algorithms (eg point,
bilinear, bicubic, and lanczos), and possibly the option of adding
sharpening too.  This can yield a significant improvement in quality.