Improvements to XML and path APIs and reorganization of Epson XML data.
The XML changes and related changes in the path API simplify the
process of loading sub-files.

The changes to the Epson driver allow for a more flexible hierarchy of
model files, which can be based on other model files (conventionally
residing in xml/escp2/model/base).

A few unused model files were removed; their model numbers can be
reused later if we want.

There are very minor (about 0.06%) changes in ink density on model 84
printers.