Store actual test logs in log directory and other testing improvements.
The automake test harness stores test logs in the tree.  During a
distcheck run, the build tree is deleted after the run, making it
impossible to review the results.  Also, each run overwrites the
previous results, which isn't conducive to the multi-step process that
scripts/build-release does.  Include the checksums in the logs, too.

Also allow specifying the log directory (via STP_LOG_DIR) to store
results externally; this would be useful for a containerized test
where the results are logged to external storage.