dyesub:  Internally rename 'laminate' to 'overcoat'
Needed because the latter is more accurate, and some printers do both.

The external option name remains 'Laminate' for backwards compatibility.

Also renamed the Magicard 'OvercoatDuplex' option to 'LaminateDuplex' so
that it's in sync with the mis-named 'Laminate' option.

In a subsequent commit I will add a new option, probably 'TrueLaminate'
for a printer that supports both Overcoat and Lamination.