dyesub:  mitsu98xx:  Support pushing gamma & matte into backend.
Code is disabled at the moment, pending testing on real hardware.

Also, wrap all non-8bpp code in the same ifdefs, so it can be nuked once
the 98xx code no longer uses it.

(That non-8bpp stuff was a ton of work.  It's a shame to see it go..)