escp2:  allow printing_initial_vertical_offset to be set via XML files
This value is used to provide a fixed vertical offset for the first
printed row on the page. It had previously been hardcoded to 0, but now
it can be specified in the model.xml file using the
<printingInitialVerticalOffset> tag.

It is specified in base_separation units, so if base_separation is 360
and the value specified is 180, this will result in the first printed
row starting 0.5 inches into the page. If not specified it will continue
to default to 0.

The (not yet committed) Epson SureLab D700 will require this feature.